Feat: 공연등록 요청 추가
